% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conesaColor.R
\name{colorConesa}
\alias{colorConesa}
\title{Return all conesa colors by number of colors you need}
\usage{
colorConesa(n, reverse = F)
}
\arguments{
\item{n}{Number of colors you need}

\item{reverse}{Boolean indicating whether the palette should be reversed}
}
\description{
Return all conesa colors by number of colors you need
}
