% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conesaColor.R
\docType{data}
\name{conesa_colors}
\alias{conesa_colors}
\title{Conesa Colors}
\format{
An object of class \code{character} of length 7.
}
\usage{
conesa_colors
}
\description{
Conesa Colors
}
\keyword{datasets}
