% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorConesa_functions.R
\name{conesa_cols}
\alias{conesa_cols}
\title{Function to extract conesa colors as hex codes}
\usage{
conesa_cols(...)
}
\arguments{
\item{...}{Character names of conesa_colors}
}
\description{
Function to extract conesa colors as hex codes
}
