% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conesaColor.R
\name{conesa_pal}
\alias{conesa_pal}
\title{Return function to interpolate a conesa color palette}
\usage{
conesa_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in conesa_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a conesa color palette
}
