% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conesaColor.R
\docType{data}
\name{conesa_palettes}
\alias{conesa_palettes}
\title{Conesa color palette}
\format{
An object of class \code{list} of length 6.
}
\usage{
conesa_palettes
}
\description{
Conesa color palette
}
\keyword{datasets}
