% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorConesa_functions.R
\name{getConesaColors}
\alias{getConesaColors}
\title{Retrieve Conesa's Main Color Set}
\usage{
getConesaColors()
}
\value{
A list containing the primary colors from the Conesa collection. Each color in the list is represented as a hexadecimal color value.
}
\description{
The \code{getConesaColors} function provides access to a curated set of colors that are part of the RColorConesa package. These colors have been specifically chosen for their utility in scientific visualizations.
}
\details{
When using the \code{getConesaColors} function, users can seamlessly integrate these colors into their R visualizations, benefiting from the expertise embedded in the Conesa color selection.
}
\examples{
getConesaColors()

}
\author{
Pedro Salguero Garcia. Maintainer: pedsalga@upv.edu.es
}
