% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorConesa_functions.R
\name{scale_color_conesa}
\alias{scale_color_conesa}
\title{Color scale constructor for conesa colors}
\usage{
scale_color_conesa(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in conesa_palettes c("main", "cool", "warm", "hot", "nature", "mixed") are avairable.}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for conesa colors
}
