% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorConesa_functions.R
\name{colorConesa}
\alias{colorConesa}
\title{colorConesa}
\usage{
colorConesa(n, reverse = F, palette = "complete")
}
\arguments{
\item{n}{Number of colors you need}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{palette}{Character name of palette in conesa_palettes: c("main", "nature", "sunshine", "hot", "warm", "cold", "complete") are avairable.}
}
\description{
Returns as many colours as you need from ConesaLab's color palettes.
}
\examples{
colorSpecies <- colorConesa(3, palette = "main")
plot(x = iris$Sepal.Length, y = iris$Sepal.Width, col = colorSpecies[iris$Species], pch = 16)

}
